Add safe guard for DB operations: check if city exists

+++ server/routes/weather.js
@@ -26,22 +26,39 @@ weather_router.route('/')
     .post(async (req, res) => {
         var city = req.body.city;
         var weather = req.body.weather;
-        await DBHandler.AddCityWeather(city, weather);
-        res.status(201);
-        res.send(`city ${city} added`);
+        var city_exists = await DBHandler.CityExists(city);
+        if (city_exists) {
+            res.status(403);
+            res.send(`city ${city} already exists, try updating its temp instead!`);
+        } else {
+            await DBHandler.AddCityWeather(city, weather);
+            res.status(201);
+            res.send(`city ${city} added`);
+        }
     })
     //Update city weather
     .put(async (req, res) => {
         var city = req.body.city;
         var weather = req.body.weather;
-        await DBHandler.UpdateCityWeather(city, weather);
-        res.send(`city ${city} updated`);
+        if (city_exists) {
+            await DBHandler.UpdateCityWeather(city, weather);
+            res.send(`city ${city} updated`);
+        } else {
+            res.status(403);
+            res.send(`city ${city} doesn't exists!`);
+        }
     })
     // Delete city
     .delete(async(req, res) => {
         var city = req.body.city;
-        await DBHandler.DeleteCity(city);
-        res.send(`city ${city} deleted`);
+        var city_exists = await DBHandler.CityExists(city);
+        if (city_exists) {
+            await DBHandler.DeleteCity(city);
+            res.send(`city ${city} deleted`);
+        } else {
+            res.status(403);
+            res.send(`city ${city} doesn't exists!`);
+        }
     });
